requirement.txt
streamlit==1.39.0
streamlit-drawable-canvas
pillow